+++ straight_skeleton/src/main/java/util/Point.java
@@ -1,14 +1,15 @@
 package at.tugraz.igi.util;
 
+import lombok.*;
 import java.util.ArrayList;
 import java.util.List;
 
 import at.tugraz.igi.main.Controller;
 
 public class Point implements Cloneable {
-	private int number;
-	private double original_x;
-	private double original_y;
+	protected int number;
+	protected double original_x;
+	protected double original_y;
 	public double current_x;
 	public double current_y;
 	public List<Line> adjacentLines;
@@ -167,6 +168,16 @@ public class Point implements Cloneable {
 		return result;
 	}
 
+	/**
+	 * We want to be able to differentiate between Points with same id ("number") and
+	 * different coordinates (e.g. for keeping snapshots history).
+	 *
+	 * Apparently some code depends on conflating those though, since adding this functionality to
+	 * the Point class breaks some test cases (which then produce skeletons with unconnected
+	 * closely placed points).
+     *
+	 * Also, it seems we need to change the behaviour globally for snapshots to work, hence the hack.
+	 */
 	@Override
 	public boolean equals(Object obj) {
 		if (this == obj)
@@ -179,9 +190,17 @@ public class Point implements Cloneable {
 		Point other = (Point) obj;
 		if (number != other.number)
 			return false;
-		if (number == other.number) {
+
+		if (!Controller.HISTORY_MODE && number == other.number)
 			return true;
-		}
+		if (number != other.number)
+			return false;
+
+		if (Controller.HISTORY_MODE && (
+		    Math.abs(original_x - other.original_x) > 1e-5 || Math.abs(original_y - other.original_y) > 1e-5 ||
+		    Math.abs(current_x - other.current_x) > 1e-5 || Math.abs(current_y - other.current_y) > 1e-5))
+			return false;
+
 		return true;
 	}