+++ straight_skeleton/src/main/java/util/StraightSkeleton.java
@@ -1,5 +1,7 @@
 package at.tugraz.igi.util;
 
+import lombok.*;
+
 import java.awt.Color;
 import java.util.ArrayList;
 import java.util.List;
@@ -7,10 +9,10 @@ import java.util.Map;
 
 public class StraightSkeleton {
 	public Map<List<Line>, MeasuringData> polygon;
-	private List<Line>lines;
-	private List<Line> polyLines;
-	private Color color;
-	private boolean visible;
+	@Getter @Setter private List<Line>lines;
+	@Getter @Setter private List<Line> polyLines;
+	@Getter @Setter private Color color;
+	@Getter @Setter private boolean visible;
 	
 	public StraightSkeleton(){
 		lines = new ArrayList<Line>();
@@ -27,34 +29,8 @@ public class StraightSkeleton {
 //		this.polyLines.clear();
 		this.lines.clear();
 	}
-	public List<Line> getLines() {
-		return lines;
-	}
-	
 	public boolean contains(Line l){
 		return lines.contains(l);
 	}
-	public Color getColor() {
-		return color;
-	}
-	public void setColor(Color color) {
-		this.color = color;
-	}
-	public List<Line> getPolyLines() {
-		return polyLines;
-	}
-	public void setPolyLines(List<Line> polyLines) {
-		this.polyLines = polyLines;
-//		this.polyLines.clear();
-//		this.polyLines.addAll(polyLines);
-	}
-	public boolean isVisible() {
-		return visible;
-	}
-	public void setVisible(boolean visible) {
-		this.visible = visible;
-	}
-
-	
 }