+++ straight_skeleton/src/main/java/util/TreeCheck.java
@@ -29,5 +29,4 @@ public class TreeCheck {
         return GraphTests.isTree(gr);
     }
 
-
 }