+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Implemented basic browsing back through events history
+* Show triangulation as first step
+* Singe/batch runners count event & check tree structure of the results
+* Polygon loading parser: ordering bugs fixed
 * Added a batch runner & a set of test cases
 * Implemented plaintext/xz-compressed coordinates reading & loading
 * Added autonomous single file runner